Add sc2_simulator depedency + remove detailed flag
sc2_simulator is now on PyPI as a package, so it's been added as a dependency to install_requires and requirements.txt. The detailed flag served as a hacky way for me to easily get relevant player info for zephyrus.gg. Now there is a to_json method that has been implemented which serves the same purpose.